removed audioplayer package

+++ lib/utils/validators.dart
@@ -1,9 +1,9 @@
-class ValidatorFunctions{
-    static String? validateInput(String? input) {
+class ValidatorFunctions {
+  static String? validateInput(String? input) {
     if (input == null || input.isEmpty) {
       return 'Please enter a value';
     }
-    final parsedValue = int.tryParse(input);
+    final parsedValue = double.tryParse(input);
     if (parsedValue == null) {
       return 'Invalid number';
     }
